Exit code should match success/failure status
This makes it easier to use it to stop CI steps, without having to re-calculate overall test results status.

In CI, `continue-on-error` can be set to `true` if trx is run as a separate step.

The backwards compatibility option `--no-exit-code` can be used to ignore failures and always return success/0.

This also means you can safely chain calls to `dotnet test` without losing the status code result, such as:

`dotnet test -l trx || trx`

this would result in a trx report only if dotnet-test failed, and preserve the failure after trx issues the report.