Quieter output by default
Most projects will have a large number of tests. In such cases, running the report will tipically be necessary in order to know if something is failing. Even skipped tests might be quite common if they are skipped depending on the current platform (for xplat matrix runs in CI, for example).

As such, the default should be something that is sensible, rather than verbose.

We introduce the new `-v|--verbosity` switch with the following options:
- quiet: only failed tests are displayed
- normal: failed and skipped tests are displayed
- verbose: failed, skipped and passed tests are displayed

The default is `quiet`. We hide the older Skipped and implement it in terms of the new property.

This also means we must remove the `-v` alias for `--version`, since it would be ambiguous.

Reverts #58
Fixes #59